check object image file validators in POST /image/add route
- convert body object position validator value to number
- create validateAllObjectImageFiles middleware
- update multerErrorHandler middleware by change condition
  statement if to switch
- import and add validateAllObjectImageFile to POST /image/add
  route
- add multer upload.fields middleware and multerErroHandle
  middleware to POST /image/add route

+++ src/middlewares/bodyValidators.ts
@@ -66,6 +66,6 @@ export const objectImageValidators = [
     'objects.*.startY',
     'objects.*.endX',
     'objects.*.endY'
-  ]).trim().notEmpty().withMessage("must not be empty").isInt({ min: 0}).withMessage('must be a number'),
+  ]).trim().notEmpty().withMessage("must not be empty").isInt({ min: 0}).withMessage('must be a number').toInt(),
   body("objects.*.filename").trim().notEmpty().withMessage("must not be empty").isLength({min: 3}).withMessage("must have at least 3 characters")
 ];
